Fix integration tests on ganache
